fix: useMemo so bandCocmap can be in useEffect deps
